add activate & suspend endpoint http

+++ src/modules/user/user.service.ts
@@ -3,6 +3,7 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 import { UserRepository } from './user.repository';
+import { UserState } from './user.types';
 
 @Injectable()
 export class UserService {
@@ -14,7 +15,7 @@ export class UserService {
     entity.email = createUserDto.email;
     entity.password = createUserDto.password;
     entity.role = createUserDto.role;
-    entity.state = createUserDto.state;
+    entity.state = UserState.CREATED;
 
     return this.userRepository.create(entity);
   }
@@ -27,10 +28,29 @@ export class UserService {
     return this.userRepository.update(entity, updateUserDto);
   }
 
-  async remove(id: number): Promise<User> {
+  async activate(id: number): Promise<User> {
     const entity = await this.userRepository.findById(id);
 
     if (!entity) throw new NotFoundException(`User with id: ${id} not found`);
+
+    return this.userRepository.update(entity, { state: UserState.ACTIVE });
+  }
+
+  async suspend(id: number): Promise<User> {
+    const entity = await this.userRepository.findById(id);
+
+    if (!entity) throw new NotFoundException(`User with id: ${id} not found`);
+
+    return this.userRepository.update(entity, { state: UserState.SUSPENDED });
+  }
+
+  async delete(id: number, force: boolean): Promise<User> {
+    const entity = await this.userRepository.findById(id);
+
+    if (!entity) throw new NotFoundException(`User with id: ${id} not found`);
+
+    if (!force)
+      return this.userRepository.update(entity, { state: UserState.DELETED });
 
     return this.userRepository.delete(entity);
   }
